BasesAndSamples.jl

BasesAndSamples.jl provides functions for generating some polynomial bases and point sets.

BasesAndSamples.basis_chebyshev — Method

Generate a basis of chebyshev polynomials up to degree d (inclusive)

BasesAndSamples.basis_gegenbauer — Method

Basis for the Gegenbauer polynomials in dimension n up to degree d. This is the Gegenbauer polynomial with parameter lambda = n/2-1, or the Jacobi polynomial with alpha = beta = (n-3)/2. Normalized to evaluate to 1 at 1. Taken from arxiv/2001.00256, ancillary files, SemidefiniteProgramming.jl

BasesAndSamples.basis_jacobi — Function

Generate the Jacobi polynomials with parameters alpha and beta up to degree d (inclusive)

BasesAndSamples.basis_laguerre — Method

Generate the Laguerre polynomials with parameter alpha up to degree d (inclusive)

BasesAndSamples.basis_monomial — Method

Generate the monomial basis in variables x... up to degree d (inclusive)

BasesAndSamples.sample_points_chebyshev — Function

Generate the d+1 chebyshev points in [a,b]

BasesAndSamples.sample_points_chebyshev_mod — Function

Generate the d+1 modified chebyshev points in [a,b]

BasesAndSamples.sample_points_padua — Method

Generate the Padua points for degree d

BasesAndSamples.sample_points_rescaled_laguerre — Method

Generate 'rescaled laguerre' points, as in SDPB

BasesAndSamples.sample_points_simplex — Method

Generate the sample points in the unit simplex with denominator d