BasesAndSamples.jl

BasesAndSamples.jl provides functions for generating some polynomial bases and point sets.

BasesAndSamples.basis_chebyshev — Method

basis_chebyshev(d::Int,x)

Generate a basis of chebyshev polynomials up to degree d (inclusive). (https://en.wikipedia.org/wiki/Chebyshev_polynomials)

BasesAndSamples.basis_gegenbauer — Method

basis_gegenbauer(d, n, x)

Basis for the Gegenbauer polynomials in dimension n up to degree d. This is the Gegenbauer polynomial with parameter lambda = n/2-1, or the Jacobi polynomial with alpha = beta = (n-3)/2. Normalized to evaluate to 1 at 1. Taken from arxiv/2001.00256, ancillary files, SemidefiniteProgramming.jl.

BasesAndSamples.basis_jacobi — Function

basis_jacobi(d::Integer, alpha, beta, x, normalized = true)

Generate the Jacobi polynomials with parameters alpha and beta up to degree d (inclusive). (https://en.wikipedia.org/wiki/Jacobi_polynomials)

BasesAndSamples.basis_laguerre — Method

basis_laguerre(d::Integer, alpha, x)

Generate the Laguerre polynomials with parameter alpha up to degree d (inclusive). (https://en.wikipedia.org/wiki/Laguerre_polynomials)

BasesAndSamples.basis_monomial — Method

basis_monomial(d::Int, x...)

Generate the monomial basis in variables x... up to degree d (inclusive).

BasesAndSamples.sample_points_chebyshev — Function

sample_points_chebyshev(d, a = -1, b = 1)

Generate the d+1 chebyshev points in [a,b]. (https://en.wikipedia.org/wiki/Chebyshev_nodes)

BasesAndSamples.sample_points_chebyshev_mod — Function

sample_points_chebyshev_mod(d, a = -1, b = 1)

Generate the d+1 modified chebyshev points in [a,b], the chebyshev points divided by cos(pi/(2(d+1))).

BasesAndSamples.sample_points_padua — Method

sample_points_padua(d)

Generate the Padua points for degree d. (https://en.wikipedia.org/wiki/Padua_points)

BasesAndSamples.sample_points_rescaled_laguerre — Method

sample_points_rescaled_laguerre(d)

Generate 'rescaled laguerre' points, as in SDPB.

BasesAndSamples.sample_points_simplex — Method

sample_points_simplex(n, d)

Generate the rational sample points in the unit simplex with denominator d.